+++ CHANGELOG.md
@@ -15,6 +15,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ### Fixed
 
+- Ref instances now cannot be created with an empty ID
+- Deployment report correctly handles missing objects
+
 ## [17.0.18] - 2026-04-08
 
 ### Added